lkl: add instant signal handling implementation
the signal handler is lazily implemented with this commit, by checking
pending signals upon the end of syscall execution and call the handler
with the syscall context (thread), which is not the proper way at all to
call signal handler.  we may need to revisit how the handler is invoked.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>