lkl: add rumpun unikernel support for LKL
This commit offers an integration with rumprun unikernel.  It is only
tested with qemu-system-x86_64 but it potentially works fine with x86
baremetal machines as well as xen hypervior.

An initial PCI device support is also implemented by relying on exposed
virtual devices from hyperviros: virtio-net device is tested for the
moment.

The rumprun repository to support lkl build is not upstreamed yet - you
can try with the following repository before it will be in.

https://github.com/libos-nuse/rumprun

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>